requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
build==1.0.3
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
ecdsa==0.18.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
packaging==23.2
passlib==1.7.4
pip-tools==7.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyproject_hooks==1.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
#uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
